Adding initialization, fixing off-by-one errors
There were two off-by-one errors so far that showed up in every
execution of the program within a few seconds.

We were not doing initialization before so all of the data structures
were empty -- it is expected by all of the methods that initialization
will be run before using the methods. (this is quite a safe assumption).