Randomized bias and weights.
Right now it seems that even randomized we are for some reason still
always producing the same outputs from the network... It should be that
with random weights.

AHA! I think I know why this is happening, the 'resize' call in
AiEngine.cpp:102 is causing the network to be biased towards always
selecting under the bias!

We need to find a better workaround than having all of those input
the value 0.0.

Notice, a quick hack is to just make the network the same height as the
input layer then the resize won't have any effect on the network.