Linear, ReLU, calculating Q(), Input normalized to (-1, 1),
Linear, ReLU Activation:
We are using linear and ReLU activation instead of perceptron now since
most literature suggests as much.

Calculating Q():
Before we were using a neural network to try and generate an action as
the output. Now we generate a set of scores as output. We now have 2**7
possible actions based on the degrees of freedom in our action.
Before we were using the neural network outputs directly to decide the
next action. Now we are using the neural network to predict the
argmax_a Q(s,a). The prior scheme was entirely incorrect, from this
scheme we can perform backward propogation.

Inputs Normalized:
We were struggling significantly to get the outputs to in some way
reflect the inputs -- i.e. for every state we were producing the same
Q() value from the neural network. We worked around this by normalizing
the State input instead of to (0, 1) to (-1, 1). This was hacked in by
just doing some simple adjustment after the clamping operation in the
AiEngine::get_input_layer function. This was the only succesful change
we made to actually have the chosen action reflected by the current
state.

Random init:
As a final note -- we update some of randomization initialization
schemes to try and workaround the issue that was resolved by Inputs
Normalization to (-1,1), we may want to go back to the original
randomization scheme.